web-service/requirements.txt
asyncpg==0.27.0
fastapi==0.89.1
uvicorn==0.20.0
requests==2.22.0
python-dotenv==0.21.0
ta==0.10.2
pandas==1.5.3
scikit-learn==1.4.0
pytest==7.2.1
httpx==0.23.3